pyproject.toml
[tool.poetry]
name = "01-nyc-taxis"
version = "0.1.2"
description = ""
authors = ["Oliver Giles <oliver.giles@lewagon.org>"]
readme = "README.md"
packages = [
    { include = "taxi_spark"},
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
pyspark = "^3.5.0"
jupyter = "^1.0.0"
apache-airflow = "^2.7.1"
psycopg2-binary = "^2.9.9"
google-cloud-storage = "^2.11.0"
python-dotenv = "^1.0.0"
apache-airflow-providers-google = "^10.10.1"
requests = "^2.32.2"
docutils = "!=0.21.post1" # issue with the pypi docutils release. Ref https://github.com/python-poetry/poetry/issues/9293

[tool.poetry.group.dev.dependencies]
ipython = "^8.0.0"
pytest = "^7.4.2"
pytest-mock = "^3.11.1"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:pkg_resources is deprecated:DeprecationWarning",
    "ignore:Deprecated call to `pkg_resources.declare_namespace:DeprecationWarning",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
